+++ lib/puppet/provider/oneview_fc_network/c7000.rb
@@ -14,20 +14,29 @@
 # limitations under the License.
 ################################################################################
 
-require File.expand_path(File.join(File.dirname(__FILE__), '..', 'login'))
-require File.expand_path(File.join(File.dirname(__FILE__), '..', 'common'))
+require_relative '../login'
+require_relative '../common'
 require 'oneview-sdk'
 
-Puppet::Type.type(:oneview_fc_network).provide(:oneview_fc_network) do
+Puppet::Type::Oneview_fc_network.provide :c7000 do
+  desc 'Provider for OneView Fiber Channel Networks using the C7000 variant of the OneView API'
+
+  confine true: login[:hardware_variant] == 'C7000'
+
   mk_resource_methods
 
   def initialize(*args)
     super(*args)
     @client = OneviewSDK::Client.new(login)
-    @resourcetype = OneviewSDK::FCNetwork
+    api_version = login[:api_version] || 200
+    @resourcetype ||= if api_version == 200
+                        OneviewSDK::API200::FCNetwork
+                      else
+                        Object.const_get("OneviewSDK::API#{api_version}::C7000::FCNetwork")
+                      end
     # Initializes the data so it is parsed only on exists and accessible throughout the methods
     # This is not set here due to the 'resources' variable not being accessible in initialize
-    @data = {}
+    @data ||= {}
   end
 
   def self.instances