+++ lib/puppet/provider/oneview_fcoe_network/c7000.rb
@@ -18,61 +18,41 @@ require_relative '../login'
 require_relative '../common'
 require 'oneview-sdk'
 
-Puppet::Type.type(:oneview_san_manager).provide(:oneview_san_manager) do
+Puppet::Type::Oneview_fcoe_network.provide :c7000 do
+  desc 'Provider for OneView Fiber Channel over Ethernet Networks using the C7000 variant of the OneView API'
+
+  confine true: login[:hardware_variant] == 'C7000'
+
   mk_resource_methods
 
   def initialize(*args)
     super(*args)
     @client = OneviewSDK::Client.new(login)
-    @resourcetype = OneviewSDK::SANManager
-    # Initializes the data so it is parsed only on exists and accessible throughout the methods
-    # This is not set here due to the 'resources' variable not being accessible in initialize
-    @data = {}
+    api_version = login[:api_version] || 200
+    @resourcetype ||= if api_version == 200
+                        OneviewSDK::API200::FCoENetwork
+                      else
+                        Object.const_get("OneviewSDK::API#{api_version}::C7000::FCoENetwork")
+                      end
+    @data ||= {}
   end
 
-  def self.instances
-    @client = OneviewSDK::Client.new(login)
-    matches = OneviewSDK::SANManager.get_all(@client)
-    matches.collect do |line|
-      name = line['name']
-      data = line.data
-      new(name: name,
-          ensure: :present,
-          data: data)
-    end
-  end
-
-  # Provider methods
   def exists?
     @data = data_parse
-    parse_provider_uri
-    pretty @data
     empty_data_check
     !@resourcetype.find_by(@client, @data).empty?
   end
 
   def create
     return true if resource_update(@data, @resourcetype)
-    @resourcetype.new(@client, @data).add
-    @property_hash[:ensure] = :present
-    @property_hash[:data] = @data
-    true
+    @resourcetype.new(@client, @data).create
   end
 
   def destroy
-    san_manager = get_single_resource_instance
-    Puppet.notice "\n\n Removing san_manager named: #{san_manager['name']}, with uri: #{san_manager['uri']}\n"
-    san_manager.remove
-    @property_hash.clear
+    get_single_resource_instance.delete
   end
 
   def found
     find_resources
-  end
-
-  def parse_provider_uri
-    return unless @data['providerUri']
-    return if @data['providerUri'].to_s[0..6].include?('/rest/')
-    @data['providerDisplayName'] = @data.delete('providerUri')
   end
 end