+++ spec/shared_context.rb
@@ -17,11 +17,8 @@
 # General context for unit testing:
 RSpec.shared_context 'shared context', a: :b do
   before :each do
-    options_120 = { url: 'https://oneview.example.com', user: 'Administrator', password: 'secret123', api_version: 120 }
-    @client_120 = OneviewSDK::Client.new(options_120)
-
-    options_200 = { url: 'https://oneview.example.com', user: 'Administrator', password: 'secret123' }
-    @client = OneviewSDK::Client.new(options_200)
+    options_api200 = { url: 'https://oneview.example.com', user: 'Administrator', password: 'secret123' }
+    @client = OneviewSDK::Client.new(options_api200)
   end
 end
 
@@ -63,8 +60,7 @@ RSpec.shared_context 'integration context', a: :b do
 
     # Create client objects:
     $config  ||= OneviewSDK::Config.load(@config_path)
-    $client_120 ||= OneviewSDK::Client.new($config.merge(api_version: 120))
-    $client     ||= OneviewSDK::Client.new($config.merge(api_version: 200))
+    $client  ||= OneviewSDK::Client.new($config.merge(api_version: 200))
   end
 
   before :each do |e|
@@ -102,8 +98,7 @@ RSpec.shared_context 'system context', a: :b do
 
     # Create client objects:
     $config  ||= OneviewSDK::Config.load(@config_path)
-    $client_120 ||= OneviewSDK::Client.new($config.merge(api_version: 120))
-    $client     ||= OneviewSDK::Client.new($config.merge(api_version: 200))
+    $client  ||= OneviewSDK::Client.new($config.merge(api_version: 200))
 
     allow_any_instance_of(OneviewSDK::Client).to receive(:appliance_api_version).and_call_original
     allow_any_instance_of(OneviewSDK::Client).to receive(:login).and_call_original